FaceParsing_Research/requirements.txt
# Core dependencies
torch==2.0.1+cu118
torchvision==0.15.2+cu118
numpy==1.24.3
opencv-python-headless==4.8.0.74
Pillow-SIMD==9.5.0
matplotlib==3.7.1
scikit-learn==1.2.2
tqdm==4.65.0
tensorboardX==2.6

# ML specific
albumentations==1.3.1
segmentation-models-pytorch==0.3.3
einops==0.6.1
timm==0.9.2
grad-cam==1.4.6

# Data processing
pandas==2.0.2
seaborn==0.12.2
h5py==3.8.0

# Development
jupyter==1.0.0
ipywidgets==8.0.6
black==23.3.0
pytest==7.3.1
pytest-cov==4.0.0

# Utilities
pyyaml==6.0
click==8.1.3
rich==13.3.5
python-dotenv==1.0.0
